# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SecuriTrento API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the SecuriTrento application",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "SecuriTrento Team",
      "url": "https://github.com/JacCons/Ingegneria_del_software_gruppo_8"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/utenti": {
      "get": {
        "summary": "Recupera tutti gli utenti",
        "tags": [
          "Utenti"
        ],
        "description": "Restituisce una lista di tutti gli utenti registrati nel sistema",
        "responses": {
          "200": {
            "description": "Lista di utenti recuperata con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Utente"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Numero totale di utenti",
                      "example": 10
                    },
                    "message": {
                      "type": "string",
                      "example": "Utenti recuperati con successo"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/utenti/{tipo}": {
      "get": {
        "summary": "Recupera utenti per tipo",
        "tags": [
          "Utenti"
        ],
        "description": "Restituisce tutti gli utenti di un tipo specifico attivi (standard, comunale, fdo)",
        "parameters": [
          {
            "in": "path",
            "name": "tipo",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "standard",
                "comunale",
                "fdo"
              ]
            },
            "description": "Tipo di utente da recuperare"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista di utenti del tipo specificato recuperata con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Utente"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "example": 5
                    },
                    "message": {
                      "type": "string",
                      "example": "Utenti di tipo standard recuperati con successo"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Tipo di utente non valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/utenti/id/{id}": {
      "get": {
        "summary": "Recupera un utente specifico tramite ID",
        "tags": [
          "Utenti"
        ],
        "description": "Restituisce i dettagli di un utente specifico identificato dall'ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID MongoDB dell'utente"
          }
        ],
        "responses": {
          "200": {
            "description": "Dettagli dell'utente recuperati con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Utente"
                    },
                    "message": {
                      "type": "string",
                      "example": "Utente recuperato con successo"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Utente non trovato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/utenti/register/{tipo}": {
      "post": {
        "summary": "Registra un nuovo utente",
        "tags": [
          "Utenti"
        ],
        "description": "Crea un nuovo utente del tipo specificato (standard, comunale, fdo)",
        "parameters": [
          {
            "in": "path",
            "name": "tipo",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "standard",
                "comunale",
                "fdo"
              ]
            },
            "description": "Tipo di utente da registrare"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UtenteRegistratoInput"
                  },
                  {
                    "$ref": "#/components/schemas/UtenteComunaleInput"
                  },
                  {
                    "$ref": "#/components/schemas/UtenteFDOInput"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utente registrato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Utente"
                    },
                    "message": {
                      "type": "string",
                      "example": "Utente standard registrato con successo"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dati non validi o utente già esistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Errore del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Utente": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID MongoDB dell'utente",
            "example": "60d21b4667d0d8992e610c85"
          },
          "nome": {
            "type": "string",
            "description": "Nome dell'utente",
            "example": "Mario"
          },
          "cognome": {
            "type": "string",
            "description": "Cognome dell'utente",
            "example": "Rossi"
          },
          "telefono": {
            "type": "string",
            "description": "Numero di telefono dell'utente",
            "example": "3401234567"
          },
          "tipoUtente": {
            "type": "string",
            "description": "Tipo di utente (discriminator)",
            "enum": [
              "UtenteRegistrato",
              "UtenteComune",
              "UtenteFDO"
            ],
            "example": "UtenteRegistrato"
          }
        }
      },
...
    }
  },
...
}